Added observable extensions for property changes
Introduced two new extension methods for INotifyPropertyChanged and INotifyPropertyChanging interfaces. These methods create an observable sequence that emits a value when the specified property changes or is changing, respectively. The current value of the property is always emitted upon subscription. Also added corresponding unit tests to verify these new functionalities. 

This is helpful for UI platforms like MAUI, Avalonia, WinUI, and WPF where UI elements will raise change notifications when a value is updated or with view model objects that often implement the INotifyPropertyChanged/ing interfaces. The advantage of this over `EveryValueChanged` on these platforms is that the properties emit and do not require inspection when the frame advances.